Finished view movie page

+++ scripts/movie_delete.js
@@ -9,14 +9,18 @@ let id = getQueryParams(queryString);
 let url = `http://localhost:3000/movies`;
 
 if (id) {
-    url += `/${id}`;
-    fetch(url, {
-        method: 'DELETE',
-        // headers: {
-        //     'Content-Type': 'application/json'
-        // },
-        // body: JSON.stringify(movie)
-    }).then(() => {
+    if (confirm('Da li ste sigurni da zelite da obrisete film?')) {
+        url += `/${id}`;
+        fetch(url, {
+            method: 'DELETE',
+            // headers: {
+            //     'Content-Type': 'application/json'
+            // },
+            // body: JSON.stringify(movie)
+        }).then(() => {
+            location.href = 'index.html';
+        })
+    } else {
         location.href = 'index.html';
-    })
+    }
 }
